If the reseacher would like to start or stop the timer at a specific time in the past (negatime timedelta), adjust the arguments BEFORE argparse's parse_args sees them such that a negative timedelta is not seen as another option: ``` $ trk start --at "-2hour" $ trk stop --at "-1hour" -m "Working on stop_at test" ```
The adjustment to the args will add a leading space such that argparse sees
" -2hour" and " -1hour" rather than "-2hour" and "-1hour",
which it misinterprets as --at not getting a value and
the --at value being another option.

So it appears that pytimeparse2 DOES support negative timedeltas.
Thank you so much @onegreyonewhite!
Love the pytimeparse2 package
https://github.com/onegreyonewhite/pytimeparse2/issues/11